requirements.txt
anyio==4.9.0
black==25.1.0
certifi==2025.1.31
cfgv==3.4.0
charset-normalizer==3.4.1
click==8.1.8
distlib==0.3.9
filelock==3.18.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
hubspot-api-client==11.1.0
identify==2.6.9
idna==3.10
mypy==1.15.0
mypy-extensions==1.0.0
nodeenv==1.9.1
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
ruff==0.11.0
six==1.17.0
sniffio==1.3.1
typing_extensions==4.12.2
urllib3==2.3.0
virtualenv==20.29.3
